refactor: changed path to SectionHeader
